Успешно пройдены тесты для UserDbStorage

+++ src/main/java/ru/yandex/practicum/filmorate/storage/user/UserDbStorage.java
@@ -73,13 +73,12 @@ public class UserDbStorage implements UserStorage {
 
     @Override
     public User get(long id) {
-        String getQuery = "SELECT * FROM users WHERE id = ?";
-        String getFriendsQuery = "SELECT f.friend_id, f.user_id, f.status FROM friends AS f WHERE f.user_id = ?";
-        User user = jdbcTemplate.queryForObject(getQuery, this::getRowMapperUser, id);
-        if (user != null) {
-
+        if (checkUserById(id)) {
+            String getQuery = "SELECT * FROM users WHERE id = ?";
+            String getFriendsQuery = "SELECT f.friend_id, f.user_id FROM friends AS f WHERE f.user_id = ?";
+            User user = jdbcTemplate.queryForObject(getQuery, this::getRowMapperUser, id);
             if (checkFriends(id)) {
-                List<Friends> friends = jdbcTemplate.query(getFriendsQuery, this::getRowMapperFriends);
+                List<Friends> friends = jdbcTemplate.query(getFriendsQuery, this::getRowMapperFriends, id);
                 List<Long> friendIds = friends.stream().map(Friends::getFriendId).collect(Collectors.toList());
                 user.setFriends(new HashSet<>(friendIds));
             } else {
@@ -99,29 +98,15 @@ public class UserDbStorage implements UserStorage {
 
     public void addFriend(long userId, long friendId) {
         if (checkUserById(userId) && checkUserById(friendId)) {
-            String addUserToFriendQuery = "INSERT INTO friends (user_id, friend_id, status) VALUES (?,?,?)";
-            String updateFriendStatus = "UPDATE friends SET status = true WHERE user_id = ? AND friend_id = ?";
+            String addUserToFriendQuery = "INSERT INTO friends (user_id, friend_id) VALUES (?,?)";
             boolean user = checkFriendsById(userId, friendId);
-            boolean friend = checkFriendsById(friendId, userId);
             if (!user) {
-                if (!friend) {
-                    jdbcTemplate.update(connection -> {
-                        PreparedStatement stmt = connection.prepareStatement(addUserToFriendQuery, new String[]{"id"});
-                        stmt.setLong(1, userId);
-                        stmt.setLong(2, friendId);
-                        stmt.setBoolean(3, false);
-                        return stmt;
-                    });
-                } else {
-                    jdbcTemplate.update(updateFriendStatus, friendId, userId);
-                    jdbcTemplate.update(connection -> {
-                        PreparedStatement stmt = connection.prepareStatement(addUserToFriendQuery, new String[]{"id"});
-                        stmt.setLong(1, userId);
-                        stmt.setLong(2, friendId);
-                        stmt.setBoolean(3, true);
-                        return stmt;
-                    });
-                }
+                jdbcTemplate.update(connection -> {
+                    PreparedStatement stmt = connection.prepareStatement(addUserToFriendQuery, new String[]{"id"});
+                    stmt.setLong(1, userId);
+                    stmt.setLong(2, friendId);
+                    return stmt;
+                });
             } else {
                 throw new RuntimeException("Заявка пользователя с id " + userId + " в друзья к пользователю с id "
                         + friendId + "уже существует!");
@@ -132,10 +117,12 @@ public class UserDbStorage implements UserStorage {
     }
 
     public void removeFriend(long userId, long friendId) {
-        String deleteFriendsQuery = "DELETE FROM friends WHERE user_id = ? AND friend_id = ?";
-        String updateFriends = "UPDATE friends SET status = false WHERE user_id = ? AND friend_id = ?";
-        jdbcTemplate.update(deleteFriendsQuery, userId, friendId);
-        jdbcTemplate.update(updateFriends, friendId, userId);
+        if (checkUserById(userId) && checkUserById(friendId)) {
+            String deleteFriendsQuery = "DELETE FROM friends WHERE user_id = ? AND friend_id = ?";
+            jdbcTemplate.update(deleteFriendsQuery, userId, friendId);
+        } else {
+            throw new UserNotFoundException("Для добавления в друзья пользователи не найдены!");
+        }
     }
 
     private boolean checkUserById(long id) {
@@ -154,7 +141,7 @@ public class UserDbStorage implements UserStorage {
     }
 
     private boolean checkFriends(long userId) {
-        String checkFriendQuery = "SELECT EXISTS(SELECT 1 FROM FRIENDS WHERE user_id = ? AND status = true)";
+        String checkFriendQuery = "SELECT EXISTS(SELECT 1 FROM FRIENDS WHERE user_id = ?)";
         return jdbcTemplate.queryForObject(checkFriendQuery, Boolean.class, userId);
     }
 
@@ -170,12 +157,9 @@ public class UserDbStorage implements UserStorage {
 
     private Friends getRowMapperFriends(ResultSet resultSet, int rowNum) throws SQLException {
         Friends friends = null;
-        if (resultSet.next()) {
-            friends = new Friends();
-            friends.setFriendId(resultSet.getLong("friend_id"));
-            friends.setUserId(resultSet.getLong("user_id"));
-            friends.setStatus(resultSet.getBoolean("status"));
-        }
+        friends = new Friends();
+        friends.setFriendId(resultSet.getLong("friend_id"));
+        friends.setUserId(resultSet.getLong("user_id"));
         return friends;
     }
 }
